Tests passing.

+++ src/main/java/uk/co/comment/document/repository/CommentRepository.java
@@ -1,15 +1,13 @@
 package uk.co.comment.document.repository;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.data.mongodb.repository.Query;
 import org.springframework.transaction.annotation.Transactional;
 import uk.co.comment.document.domain.Comment;
-import uk.co.comment.document.rest.CommentDTO;
 
-import java.util.List;
+import java.math.BigInteger;
 
 @Transactional
-public interface CommentRepository extends MongoRepository<Comment, Long> {
+public interface CommentRepository extends MongoRepository<Comment, BigInteger> {
     
 }
